backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
google-generativeai>=0.3.2
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.26.0
tqdm>=4.66.1
python-dotenv>=1.0.0
pydantic>=2.5.0
email-validator>=2.0.0

